NP-1715 Move TextField to own component
